requirements.txt
pandas
numpy
matplotlib
seaborn
pyyaml
streamlit
sqlalchemy
